freer: Eff1 -> Eff
