+++ extensible-effects.cabal
@@ -63,35 +63,27 @@ library
   -- Modules exported by the library.
   exposed-modules:     Control.Eff
                        Control.Eff.Choose
-                       Control.Eff.Choose1
                        Control.Eff.Coroutine
-                       Control.Eff.Cut
+                       -- Control.Eff.Cut
                        -- Control.Eff.Cut1
                        Control.Eff.Example
-                       Control.Eff.Example1
                        Control.Eff.Exception
-                       Control.Eff.Exception1
                        Control.Eff.Fresh
                        Control.Eff.Lift
-                       Control.Eff.Lift1
                        Control.Eff.NdetEff
-                       Control.Eff.Reader.Lazy1
-                       Control.Eff.Reader.Strict1
+                       Control.Eff.Reader.Lazy
+                       Control.Eff.Reader.Strict
+                       Control.Eff.State.LazyState
                        Control.Eff.State.Lazy
-                       Control.Eff.State.Lazy1
                        Control.Eff.State.Strict
-                       Control.Eff.State.Strict1
                        Control.Eff.Writer.Lazy
-                       Control.Eff.Writer.Lazy1
                        Control.Eff.Writer.Strict
-                       Control.Eff.Writer.Strict1
                        Control.Eff.Trace
                        Control.Eff.Operational
                        Control.Eff.Operational.Example
                        Control.Monad.Free.Reflection
                        Data.OpenUnion
                        Data.OpenUnion51
-                       Control.Eff1
                        Data.OpenUnion.Imports
 
   -- Modules included in this library but not exported.
