+++ src/Control/Eff/Reader/Strict.hs
@@ -1,17 +1,18 @@
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE NoMonomorphismRestriction #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TypeOperators #-}
--- | Lazy read-only state
-module Control.Eff.Reader.Lazy1( Reader (..)
+-- | Strict read-only state
+module Control.Eff.Reader.Strict ( Reader (..)
                               , ask
                               , local
                               , reader
                               , runReader
                               ) where
 
-import Control.Eff1 hiding (Reader(..), ask, local, runReader, runReader')
+import Control.Eff
 import Data.OpenUnion51
 import Data.FTCQueue1
 
@@ -42,7 +43,7 @@ ask = send $ Reader id
 -- | The handler of Reader requests. The return type shows that all Reader
 -- requests are fully handled.
 runReader :: Eff (Reader e ': r) w -> e -> Eff r w
-runReader m e = handle_relay
+runReader m !e = handle_relay
   return
   (\(Reader f) k -> k (f e))
   m
