+++ src/Control/Eff/State/Strict.hs
@@ -1,70 +1,123 @@
-{-# LANGUAGE TypeOperators #-}
+{-# OPTIONS_GHC -Werror #-}
 {-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE DeriveFunctor #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE Safe #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE NoMonomorphismRestriction #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeOperators #-}
 -- | Strict state effect
---
--- Example: implementing `Control.Eff.Fresh`
---
--- > runFresh' :: (Typeable i, Enum i, Num i) => Eff (Fresh i :> r) w -> i -> Eff r w
--- > runFresh' m s = fst <$> runState s (loop $ admin m)
--- >  where
--- >   loop (Val x) = return x
--- >   loop (E u)   = case decomp u of
--- >     Right (Fresh k) -> do
--- >                       n <- get
--- >                       put (n + 1)
--- >                       loop (k n)
--- >     Left u' -> send (\k -> unsafeReUnion $ k <$> u') >>= loop
-module Control.Eff.State.Strict( State (..)
-                               , get
-                               , put
-                               , modify
-                               , runState
-                               , evalState
-                               , execState
-                               ) where
-
-import Data.Typeable
+module Control.Eff.State.Strict where
 
 import Control.Eff
+import Control.Eff.Writer.Strict
+import Control.Eff.Reader.Strict
+import Data.OpenUnion51
+import Data.FTCQueue1
 
--- | Strict state effect
-data State s w = State (s -> s) (s -> w)
-  deriving (Typeable, Functor)
+-- ------------------------------------------------------------------------
+-- | State, strict
+--
+-- Initial design:
+-- The state request carries with it the state mutator function
+-- We can use this request both for mutating and getting the state.
+-- But see below for a better design!
+--
+-- > data State s v where
+-- >   State :: (s->s) -> State s s
+--
+-- In this old design, we have assumed that the dominant operation is
+-- modify. Perhaps this is not wise. Often, the reader is most nominant.
+--
+-- See also below, for decomposing the State into Reader and Writer!
+--
+-- The conventional design of State
+data State s v where
+  Get :: State s s
+  Put :: !s -> State s ()
+
+-- | Return the current value of the state. The signatures are inferred
+{-# NOINLINE get #-}
+get :: Member (State s) r => Eff r s
+get = send Get
+{-# RULES
+  "get/bind" forall k. get >>= k = send Get >>= k
+ #-}
 
 -- | Write a new value of the state.
-put :: (Typeable e, Member (State e) r) => e -> Eff r ()
-put !s = modify $ const s
+{-# NOINLINE put #-}
+put :: Member (State s) r => s -> Eff r ()
+put !s = send (Put s)
+{-# RULES
+  "put/bind"     forall k v. put v >>= k = send (Put v) >>= k
+ #-}
+{-# RULES
+  "put/semibind" forall k v. put v >>  k = send (Put v) >>= (\() -> k)
+ #-}
+-- The purpose of the rules is to expose send, which is then being
+-- fuzed by the send/bind rule. The send/bind rule is very profitable!
+-- These rules are essentially inlining of get/put. Somehow GHC does not
+-- inline get/put, even if I put the INLINE directives and play with phases.
+-- (Inlining works if I use 'inline' explicitly).
 
--- | Return the current value of the state.
-get :: (Typeable e, Member (State e) r) => Eff r e
-get = send . inj $ State id id
+runState' :: Eff (State s ': r) w -> s -> Eff r (w,s)
+runState' m !s =
+  handle_relay_s s (\s0 x -> return (x,s0))
+                   (\s0 sreq k -> case sreq of
+                       Get    -> k s0 s0
+                       Put s1 -> k s1 ())
+                   m
+
+-- Since State is so frequently used, we optimize it a bit
+-- | Run a State effect
+runState :: Eff (State s ': r) w  -- ^ Effect incorporating State
+         -> s                     -- ^ Initial state
+         -> Eff r (w,s)           -- ^ Effect containing final state and a return value
+runState (Val x) !s = return (x,s)
+runState (E u q) !s = case decomp u of
+  Right Get     -> runState (qApp q s) s
+  Right (Put s1) -> runState (qApp q ()) s1
+  Left  u1 -> E u1 (tsingleton (\x -> runState (qApp q x) s))
 
 -- | Transform the state with a function.
-modify :: (Typeable s, Member (State s) r) => (s -> s) -> Eff r ()
-modify f = send . inj $ State f $ const ()
-
--- | Run a State effect.
-runState :: Typeable s
-         => s                     -- ^ Initial state
-         -> Eff (State s :> r) w  -- ^ Effect incorporating State
-         -> Eff r (s, w)          -- ^ Effect containing final state and a return value
-runState = loop
-  where
-    loop !s = freeMap
-              (\x -> return (s, x))
-              (\u -> handleRelay u (loop s) $
-                     \(State t k) -> let s' = t s
-                                     in loop s' (k s'))
+modify :: (Member (State s) r) => (s -> s) -> Eff r ()
+modify f = get >>= put . f
 
 -- | Run a State effect, discarding the final state.
-evalState :: Typeable s => s -> Eff (State s :> r) w -> Eff r w
-evalState s = fmap snd . runState s
+evalState :: Eff (State s ': r) w -> s -> Eff r w
+evalState m !s = fmap fst . flip runState s $ m
 
 -- | Run a State effect and return the final state.
-execState :: Typeable s => s -> Eff (State s :> r) w -> Eff r s
-execState s = fmap fst . runState s
+execState :: Eff (State s ': r) w -> s -> Eff r s
+execState m !s = fmap snd . flip runState s $ m
+
+-- | An encapsulated State handler, for transactional semantics
+-- The global state is updated only if the transactionState finished
+-- successfully
+data ProxyState s = ProxyState
+transactionState :: forall s r w. Member (State s) r =>
+                    ProxyState s -> Eff r w -> Eff r w
+transactionState _ m = do s <- get; loop s m
+ where
+   loop :: s -> Eff r w -> Eff r w
+   loop s (Val x) = put s >> return x
+   loop s (E (u::Union r b) q) = case prj u :: Maybe (State s b) of
+     Just Get      -> loop s (qApp q s)
+     Just (Put s') -> loop s'(qApp q ())
+     _      -> E u (tsingleton k) where k = qComp q (loop s)
+
+-- | A different representation of State: decomposing State into mutation
+-- (Writer) and Reading. We don't define any new effects: we just handle the
+-- existing ones.  Thus we define a handler for two effects together.
+runStateR :: Eff (Writer s ': Reader s ': r) w -> s -> Eff r (w,s)
+runStateR m !s = loop s m
+ where
+   loop :: s -> Eff (Writer s ': Reader s ': r) w -> Eff r (w,s)
+   loop s0 (Val x) = return (x,s0)
+   loop s0 (E u q) = case decomp u of
+     Right (Writer w v) -> k w v
+     Left  u1  -> case decomp u1 of
+       Right (Reader f) -> k s0 (f s0)
+       Left u2 -> E u2 (tsingleton (k s0))
+    where k x = qComp q (loop x)
