+++ src/Control/Eff/Writer/Lazy.hs
@@ -1,61 +1,69 @@
-{-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE DeriveFunctor #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE Safe #-}
--- | Lazy write-only state.
-module Control.Eff.Writer.Lazy( Writer (..)
-                              , tell
-                              , censor
-                              , runWriter
-                              , runFirstWriter
-                              , runLastWriter
-                              , runMonoidWriter
-                              ) where
-
-import Data.Typeable
-import Control.Applicative ((<|>))
-import Data.Monoid
+{-# LANGUAGE NoMonomorphismRestriction #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeOperators #-}
+-- | Lazy write-only state
+module Control.Eff.Writer.Lazy ( Writer(..)
+                               , tell
+                               , censor
+                               , runWriter
+                               , runFirstWriter
+                               , runLastWriter
+                               , runListWriter
+                               , runMonoidWriter
+                               ) where
 
 import Control.Eff
+import Data.OpenUnion51
+import Data.FTCQueue1
 
--- | The request to remember a value of type w in the current environment
+import Data.Monoid
+import Control.Applicative ((<|>))
+
+-- ------------------------------------------------------------------------
+-- | The Writer monad
+--
+-- In MTL's Writer monad, the told value must have a |Monoid| type. Our
+-- writer has no such constraints. If we write a |Writer|-like
+-- interpreter to accumulate the told values in a monoid, it will have
+-- the |Monoid w| constraint then
 data Writer w v = Writer w v
-    deriving (Typeable, Functor)
 
 -- | Write a new value.
-tell :: (Typeable w, Member (Writer w) r) => w -> Eff r ()
-tell w = send . inj $ Writer w ()
+tell :: Member (Writer w) r => w -> Eff r ()
+tell w = send $ Writer w ()
 
 -- | Transform the state being produced.
-censor :: (Typeable w, Member (Writer w) r) => (w -> w) -> Eff r a -> Eff r a
-censor f = loop
+censor :: forall w a r. Member (Writer w) r => (w -> w) -> Eff r a -> Eff r a
+censor f = interpose return h
   where
-    loop = freeMap
-           return
-           (\u -> interpose u loop
-                  $ \(Writer w v) -> tell (f w) >> loop v)
+    h (Writer w v) k = tell (f w) >> k v
 
--- | Handle Writer requests, using a user-provided function to accumulate values.
-runWriter :: Typeable w => (w -> b -> b) -> b -> Eff (Writer w :> r) a -> Eff r (b, a)
-runWriter accum b = loop
-  where
-    first f (x, y) = (f x, y)
 
-    loop = freeMap
-           (\x -> return (b, x))
-           (\u -> handleRelay u loop
-                  $ \(Writer w v) -> first (accum w) `fmap` loop v)
+-- | Handle Writer requests, using a user-provided function to accumulate
+-- values, hence no Monoid constraints.
+runWriter :: (w -> b -> b) -> b -> Eff (Writer w ': r) a -> Eff r (a, b)
+runWriter accum b = handle_relay
+  (\x -> return (x, b))
+  (\(Writer w v) k -> k v >>= \(x, l) -> return (x, w `accum` l))
+  -- the second arg to 'handle_relay' above is same as:
+  -- (\(Writer w v) k -> second (accum w) `fmap` k v)
+  -- where
+  --   second f (x, y) = (x, f y)
+
+-- | Handle Writer requests, using a List to accumulate values.
+runListWriter :: Eff (Writer w ': r) a -> Eff r (a,[w])
+runListWriter = runWriter (:) []
+
+-- | Handle Writer requests, using a Monoid instance to accumulate values.
+runMonoidWriter :: (Monoid w) => Eff (Writer w ': r) a -> Eff r (a, w)
+runMonoidWriter = runWriter (<>) mempty
 
 -- | Handle Writer requests by taking the first value provided.
-runFirstWriter :: Typeable w => Eff (Writer w :> r) a -> Eff r (Maybe w, a)
+runFirstWriter :: Eff (Writer w ': r) a -> Eff r (a, Maybe w)
 runFirstWriter = runWriter (\w b -> Just w <|> b) Nothing
 
 -- | Handle Writer requests by overwriting previous values.
-runLastWriter :: Typeable w => Eff (Writer w :> r) a -> Eff r (Maybe w, a)
+runLastWriter :: Eff (Writer w ': r) a -> Eff r (a, Maybe w)
 runLastWriter = runWriter (\w b -> b <|> Just w) Nothing
-
--- | Handle Writer requests, using a Monoid instance to accumulate values.
-runMonoidWriter :: (Monoid w, Typeable w) => Eff (Writer w :> r) a -> Eff r (w, a)
-runMonoidWriter = runWriter (<>) mempty
